requirements.txt
pillow>=9.0.0
tkinter>=8.6; python_version < '3.7'